did some refactoring before pull
