+++ README.md
@@ -6,6 +6,7 @@
         - DB_CONNECTION - тут система управления БД (mysql, mssql, postgre и т.д)
         - DB_HOST - айпи на котором хостится БД. Это нужно смотерть в настройках сервера БД. Например,в OSPanel 6 mySQL 8.2 хостится на 127.127.126.50
         - DB_PORT - стандартный порт это 3306, если БД не отвечает, то лучше проверить.
+        DB_DATABASE=имя БД
 
     - устанавливаем зависимости командами
         composer install
