+++ resources/js/Components/Inputs/InputCustom/features/handleValue.ts
@@ -1,0 +1,5 @@
+function handleValue(value: Array<string> | string): string {
+    return Array.isArray(value) ? value.join(' ') : value
+}
+
+export default handleValue;
