+++ resources/js/Components/Inputs/InputCustom/features/split.ts
@@ -1,0 +1,5 @@
+function split(value: string): Array<string> {
+    return value.split(/[\s,;]+/)
+}
+
+export default split;
