+++ resources/js/Layouts/Main.tsx
@@ -1,7 +1,7 @@
 import { Header } from '@/Components/Header'
 import { MainButton } from '@/Components/Buttons/MainButton'
 import { TableSearchingForm } from '@/Components/TableSerachingForm/TableSearchingForm'
-import { Table } from '@/Components/Table'
+import { Table } from '@/Components/Table/Table'
 import { Toggle } from '@/Components/Buttons/Toggle'
 import { DevTool } from '@hookform/devtools'
 import { useFormContext } from 'react-hook-form'
